refactor: Update delete endpoint to handle optional user ID

+++ src/main/java/com/mycompany/SkySong/authentication/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.mycompany.SkySong.authentication.controller;
 
+import com.mycompany.SkySong.authentication.exception.NullOrEmptyInputException;
 import com.mycompany.SkySong.authentication.model.dto.DeleteResponse;
 import com.mycompany.SkySong.authentication.service.DeleteService;
 import org.springframework.http.HttpStatus;
@@ -17,8 +18,11 @@ public class UserController {
     public UserController(DeleteService deleteService) {
         this.deleteService = deleteService;
     }
-    @DeleteMapping("/{userId}")
-    public ResponseEntity<DeleteResponse> delete(@PathVariable Long userId) {
+    @DeleteMapping({"/","/{userId}"})
+    public ResponseEntity<DeleteResponse> delete(@PathVariable(required = false) Long userId) {
+        if (userId == null) {
+            throw new NullOrEmptyInputException("User ID cannot be empty");
+        }
         DeleteResponse deleteResponse = deleteService.deleteUser(userId);
         return new ResponseEntity<>(deleteResponse, HttpStatus.OK);
     }
